+++ sae-20190506/src/Models/ListApplicationsResponseBody/data/applications/children.php
@@ -90,6 +90,11 @@ class children extends Model
     public $regionId;
 
     /**
+     * @var string
+     */
+    public $resourceType;
+
+    /**
      * @var int
      */
     public $runningInstances;
@@ -125,6 +130,7 @@ class children extends Model
         'newSaeVersion' => 'NewSaeVersion',
         'programmingLanguage' => 'ProgrammingLanguage',
         'regionId' => 'RegionId',
+        'resourceType' => 'ResourceType',
         'runningInstances' => 'RunningInstances',
         'scaleRuleEnabled' => 'ScaleRuleEnabled',
         'scaleRuleType' => 'ScaleRuleType',
@@ -204,6 +210,10 @@ class children extends Model
 
         if (null !== $this->regionId) {
             $res['RegionId'] = $this->regionId;
+        }
+
+        if (null !== $this->resourceType) {
+            $res['ResourceType'] = $this->resourceType;
         }
 
         if (null !== $this->runningInstances) {
@@ -302,6 +312,10 @@ class children extends Model
 
         if (isset($map['RegionId'])) {
             $model->regionId = $map['RegionId'];
+        }
+
+        if (isset($map['ResourceType'])) {
+            $model->resourceType = $map['ResourceType'];
         }
 
         if (isset($map['RunningInstances'])) {