+++ sae-20190506/src/Models/ResumeTrafficResponseBody/data.php
@@ -1,0 +1,62 @@
+<?php
+
+// This file is auto-generated, don't edit it. Thanks.
+
+namespace AlibabaCloud\SDK\Sae\V20190506\Models\ResumeTrafficResponseBody;
+
+use AlibabaCloud\Dara\Model;
+
+class data extends Model
+{
+    /**
+     * @var string
+     */
+    public $msg;
+
+    /**
+     * @var bool
+     */
+    public $success;
+    protected $_name = [
+        'msg' => 'msg',
+        'success' => 'success',
+    ];
+
+    public function validate()
+    {
+        parent::validate();
+    }
+
+    public function toArray($noStream = false)
+    {
+        $res = [];
+        if (null !== $this->msg) {
+            $res['msg'] = $this->msg;
+        }
+
+        if (null !== $this->success) {
+            $res['success'] = $this->success;
+        }
+
+        return $res;
+    }
+
+    public function toMap($noStream = false)
+    {
+        return $this->toArray($noStream);
+    }
+
+    public static function fromMap($map = [])
+    {
+        $model = new self();
+        if (isset($map['msg'])) {
+            $model->msg = $map['msg'];
+        }
+
+        if (isset($map['success'])) {
+            $model->success = $map['success'];
+        }
+
+        return $model;
+    }
+}